Split utils in ros, bt and normal ones
