+++ jani_generator/src/jani_generator/ros_helpers/ros_services.py
@@ -24,11 +24,10 @@ from scxml_converter.scxml_entries import (ScxmlAssign, ScxmlData,
                                            ScxmlDataModel, ScxmlParam,
                                            ScxmlRoot, ScxmlSend, ScxmlState,
                                            ScxmlTransition)
-from scxml_converter.scxml_entries.utils import (
-    generate_srv_request_event, generate_srv_response_event,
-    generate_srv_server_request_event, generate_srv_server_response_event,
-    get_default_expression_for_type, get_srv_type_params,
-    sanitize_ros_interface_name)
+from scxml_converter.scxml_entries.utils import get_default_expression_for_type
+from scxml_converter.scxml_entries.ros_utils import (
+    generate_srv_request_event, generate_srv_response_event, generate_srv_server_request_event,
+    generate_srv_server_response_event, get_srv_type_params, sanitize_ros_interface_name)
 
 SRV_PREFIX = "srv_handler_"
