+++ scxml_converter/src/scxml_converter/scxml_entries/scxml_ros_field.py
@@ -53,8 +53,7 @@ class RosField(ScxmlParam):
         return valid_name and valid_expr
 
     def as_plain_scxml(self, _) -> ScxmlParam:
-        from scxml_converter.scxml_entries.utils import \
-            replace_ros_interface_expression
+        from scxml_converter.scxml_entries.ros_utils import replace_ros_interface_expression
         return ScxmlParam(self._name, expr=replace_ros_interface_expression(self._expr))
 
     def as_xml(self) -> ET.Element:
