+++ scxml_converter/src/scxml_converter/scxml_entries/scxml_ros_service.py
@@ -23,13 +23,11 @@ https://docs.ros.org/en/iron/Tutorials/Beginner-CLI-Tools/Understanding-ROS2-Ser
 from typing import List, Optional, Union
 from xml.etree import ElementTree as ET
 
-from scxml_converter.scxml_entries import (RosField, ScxmlBase,
-                                           ScxmlExecutionBody, ScxmlSend,
-                                           ScxmlTransition,
-                                           as_plain_execution_body,
-                                           execution_body_from_xml,
-                                           valid_execution_body)
-from scxml_converter.scxml_entries.utils import (
+from scxml_converter.scxml_entries import (
+    RosField, ScxmlBase, ScxmlExecutionBody, ScxmlSend, ScxmlTransition,
+    as_plain_execution_body, execution_body_from_xml, valid_execution_body)
+
+from scxml_converter.scxml_entries.ros_utils import (
     ScxmlRosDeclarationsContainer, generate_srv_request_event,
     generate_srv_response_event, generate_srv_server_request_event,
     generate_srv_server_response_event, is_srv_type_known)
